Bug #33930766 hypergraph: Assert failure in ApplyWindowFunctions() in query with 64 windows
A SELECT query with 64 window functions crashes in hypergraph while
processing the 64th window when all the ORDER BY clauses are on
different indices.

1) Crash in SELECT while applying windows:
The hypergraph optimizer looks up the interesting-orders list to
find the order which will match the maximum possible windows, and
returns an index into the interesting-order list. This index is
used to index into the interesting-order list to create a sort
access path to process the selected windows. This is repeated until
all the windows are processed. The problem occurs when the lookup
fails because all the remaining windows have indices which exceed
the maximum interesting-order index, so none of them can match any
interesting order. This is expected never to happen and results in
an assertion failure.

The bug is fixed by first processing all the windows which match an
interesting order, and then handling the remaining windows one by
one. For each remaining window, a sort path is created using the
window's own ordering index so that the windows can be applied. The
remaining window orders are not considered as interesting orders,
which is considered acceptable since 64 or more window functions is
a corner case.

2) Crash in SELECT DISTINCT while applying DISTINCT
While handling DISTINCT, all the sort-ahead orderings are checked
against the DISTINCT ordering index, and if a match is found, a
sort path is created. While comparing indices, matching indices are
missed if the index is greater than kMaxSupportedOrderings, even
though the orderings list extends beyond kMaxSupportedOrderings. Due
to this, no matching index is returned for a distinct index which is
greater than kMaxSupportedOrderings, and no access paths are
created for it, causing a crash.

This bug is fixed by creating a sort access path using the distinct
ordering index if no matching index is found.

Change-Id: I6a5eac3471d08ef1c0d6fa971d579f71f593aa0d