Bug#34063499: Hypergraph Result mismatch : NULL SAFE EQUAL [optimizer]
Enable optimizer support for creating hash joins that have equijoin
predicates with the <=> operator (NULL-safe equal). This is enabled
both in the old optimizer and in the hypergraph optimizer.

It is enabled by making Item_func::contains_only_equi_join_condition()
recognize Item_func_equal as an equijoin condition in the same way as
Item_func_eq.

Change-Id: I7cb8103ba2feae681d478ea5d77c9845e294d054