BUG#32688091: TIMEOUT ON GR_EXIT_STATE_ACTION_ON_* TESTS
On tests meant to validate Group Replication behaviour when errors
happen during START GROUP_REPLICATION - like incompatible
configuration, on which case the new member leaves the group
immediately, strange deadlocks were being observed on Windows.
After detailed analysis the deadlock was caused by:
  Thread 1: does a readlock on X
  Thread 2: waits for a writelock on X
  Thread 1: waits for a readlock on X
the second readlock from Thread 1 is not allowed to avoid
starvation, but Thread 2 cannot proceed since Thread 1 does not
release the first lock.

Although the scenario is simple to understand, it can happen when
the locks are hidden under a API.
This specific scenario happens in interactions with the service
infrastructure API.
Thread 1 is a notifier service that delivers status and membership
to listeners that register on `group_member_status_listener` and
`group_membership_listener` respectively.
Thread 1 was creating a service listeners iterator, which
acquires and holds a read-lock over the services registry, and then
does acquire a second read-lock so that it can notify each listener.
Thread 2 was the member unregistering a service, which does acquire
a write-lock on the services registry.

To avoid the above issue, the code on thread 1 was changed to:
 1) create the iterator;
 2) store all listener services names on a local list;
 3) release the iterator;
 4) walk over the local list:
    a) acquire the listener service;
    b) call the listener service;
    c) release the listener service;
ensuring that Thread 1 only acquires a read-lock at a time.

Change-Id: I9c3c6d3abbfb17abebbff954c2c7cb65c0475580