Bug#33640758: Assertion failed when changing value of binlog_expire_logs_seconds
Bug#26274274 tries to handle the case of both
binlog_expire_logs_seconds and the deprecated
expire_logs_days being set.

It uses the same message regardless of whether
the assignment happens at start-up (when the
message must be printed to the error-log, as
no session is attached), and when doing it
dynamically in a SQL session. This let to an
assertion.

Additionally, the behavior is different in each
case (at start-up, a warning is thrown and
expire_logs_days is ignored; at run-time, an
error is thrown and the assignment to
binlog_expire_logs_seconds fails.

This changeset adds a separate message for
the interactive case. This both fixes the
assertion and lets us use a dedicated message
that does not incorrectly describe the
goings-on (like using the start-up message
for the interactive case did).

Since the code and the message disagreed,
the original commit-message and the official
MySQL documentation were used to deduced the
original intent.

A suitable testlet is included.

Change-Id: I461849869bd3b67fe7a00e7b408b1b81c101dfda