Bug #34094271 Combinations files cannot contain options with values which evaluate to false
Issue:
======
If a combination file has an option that takes a numeric value and
one of the combinations uses the value 0 as the option's argument,
server start will fail complaining that the option is being used
without a value.

Fix:
====
Change the condition for assigning values to an option from if ({value})
to if (defined {value}).

Change-Id: I275109a05ccce4645a319cd97d03ed71d0c77ae1