Bug#34131402: uint Filesort::make_sortorder(ORDER *, bool): Assertion `count > 0' failed.
An assert failure was seen in some aggregated queries that had an
ORDER BY clause that was made redundant by an equality predicate, and
the equality predicate could not be pushed down as a table
predicate. (In practice, this means the predicate had to be an IS NULL
predicate on a column in the inner table of an outer join, and the
ORDER BY clause referenced the same column.)

The assert failure was raised because the query plan contained a SORT
path for the ORDER BY clause, but the sort key of that path was empty.

The SORT path was added because LogicalOrderings::DoesFollowOrder(),
when called in ApplyDistinctAndOrder(), was not capable of seeing that
the AGGREGATE path already had the order required by the ORDER BY
clause. This again happens because, during construction of the state
machines in LogicalOrderings, it is assumed that a grouping never
needs to become an ordering. And it's true that, in general, a
grouping cannot become an ordering without sorting. It is however
possible that a grouping satisfies an ordering when there is an
equality predicate that ensures one of the columns have only one
specific value.

LogicalOrderings is usually capable of seeing such trivial orderings,
even after grouping, because these predicates tend to be pushed down
as table predicates which result in always active functional
dependencies in the interesting orderings machinery. Orderings that
come from always active functional dependencies are baked into the
initial state of the resulting DFSM. The problem here was that the
predicate could not be pushed down, and its corresponding functional
dependency is therefore not always active.

The assert failure is fixed by not adding a SORT node when we have
found that the entire ORDER BY clause is redundant, and not even
checking what DoesFollowOrder() says.

One might argue that a more correct fix would have been to improve the
state machine in LogicalOrderings so that DoesFollowOrder() doesn't
give a false negative in this case. And it is entirely possible to add
edges from the interesting groupings to these trivial orderings in the
state machine in order to get the correct answer. It is not done in
this patch because:

1) It's not immediately obvious that it would have any benefits for
other than corner cases, so it's better to keep the complexity of the
state machine lower. At least until we find a case where the benefit
is more obvious.

2) The extra check added in this patch is needed even if we improve
the state machine. This is because the size of the state machine is
capped (see kMaxNFSMStates and kMaxDFSMStates) to avoid runaway
computation if the query is very complex. We must therefore always
account for the possibility of a false negative answer from
DoesFollowOrder(), simply because the state machine might be
incomplete.

Change-Id: Ie81b4ad6e7b69766c982ccf628282eb6fbdbcca3