Bug#34224193 Adding default value for not null column with ALTER TABLE have issues.
NDB does not support inplace change of the default
value for a column, it can only be done using copying alter
table. Changing of the default value value was already
detected, but not adding or removing default value from a column.

Fix by detecting when default value is added or removed
during ALTER TABLE and refuse to do it inplace.

Change-Id: Ica06b91dec7d7631e0027096cc2878896328a54e