Bug #34231639: compile error with the newest Visual studio 2022
Visual studio pegs the value of __cplusplus to some old pre-C++11
default. It can do proper values but this has to be turned on
specifically.

Thus remove the comparison for the __cplusplus value and just rely
that all C++ compilers are going to support
C++11 (this is checked elsewhere)

backport to mysql-8.0

Change-Id: Ic632524f541e81846123f4284e84e3979ae1c390