Bug#33535746: Cursor is not reset for consecutive prepared statement executions
Some connectors are capable of executing prepared statements alternately
with and without a cursor. This feature was removed in WL#9384, because
it was assumed that no connector would do such thing. Before WL#9384,
special support was not needed, since each execution was prepared
individually. However, with prepared once semantics, we actually need to
handle this situation.

The solution chosen is creating and saving two query result objects,
one for use with a cursor and one for use without a cursor. The objects
are created on demand, thus if a prepared statement is never used with
cursors, the cursor-specific query result object is never created.
(The non-cursor-specific query result is always created, though, since
it is created by Prepared_statement::prepare()).

This support demands significant changes in the Prepared_statement
class and cursor handling, thus we have also taken the opportunity to
perform some refactoring of the class and of prepared statement and
cursor execution. These changes are summarized below:

class Prepared_statement

- the IN_USE and IS_SQL_PREPARE flags are replaced with boolean member
  variables.

- the thd member is removed, since all member functions have access to
  a supplied thd argument anyway (however some internal functions had
  to be extended with a thd argument).

- Quite a few member variables have been adjusted to adhere to the
  current naming style, by prefixing the name with "m_".

- A new member m_used_as_cursor has been added. It is true when the
  execution is with a cursor and is used to indicate whether closing
  the cursor from a prior execution is needed.

- The member "result" is replaced with two new ones: m_regular_result
  and m_cursor_result, to keep track of the query result objects used
  for non-cursor-based and cursor-based execution, respectively.

The function Prepared_statement::execute() has been rewritten quite
a bit:

- A scope guard has been implemented to ensure proper cleanup both
  after successful and unsuccessful execution. This means that less
  state has to be managed inside the function.

- Resource group handling has been generalized so that it also works
  with cursor-based statements. Earlier it was effective only with
  regular execution.

- On first execution with and without cursor, the appropriate query
  result object is created or retrieved and stored in the
  Prepared_statement object.

- When used with a cursor, the mysql_open_cursor() function is no
  longer called. Instead, mysql_execute_command() executes the statement
  both with and without cursor. To achieve this, a new function
  new_cursor_result() is implemented that creates a cursor-specific
  query result object when needed.

- When used with a cursor, we also need to open the cursor after
  execution.

The function mysql_open_cursor() is no longer used to execute prepared
statements, but is still needed for cursors that are created within
stored procedures. Functionality that was only required for prepared
statements has been removed.

Miscellaneous changes:

- The Query_result function check_simple_query_block() has been
  renamed to check_supports_cursor(), since this is what it is used for.

- check_simple_query_block() has been created for class
  Query_result_to_file (same implementation as for Query_dumpvar).

- The Query_result class has a new member function cursor() that returns
  a cursor associated with the object, if any.

Finally, a new test that alternates between using a prepared statement
with and without a cursor has been implemented.

Change-Id: I1966d0a4059a67b7453d3dc208905a8f34c8cf0b