Bug#34190385: Hypergraph : Support for EXPLAIN or at least mention that Using secondary engine [3/3, fix]
Add rudimentary support for EXPLAIN with tabular format when using the
hypergraph optimizer with queries that use a secondary storage engine.
The tabular format is not currently supported when using the
hypergraph optimizer, so it only shows a single line saying that a
secondary storage engine is used and that EXPLAIN FORMAT=TREE should
be used to show the plan.

Since the EXPLAIN statements no longer give an error when running with
the hypergraph optimizer, "mtr --hypergraph" no longer automatically
filters out the difference between the old optimizer and the
hypergraph optimizer. Those tests in the secondary_engine test suite
that test EXPLAIN statements are therefore split into two tests; one
that runs with the old optimizer and one that runs with the hypergraph
optimizer.

Change-Id: I14832179989ee20325d4d8441927ec1143accf8c