Bug#34122738 Write_rows apply failure if source table has index and replica an extra PK
Problem
-------

If in the replica, a table has an extra primary key when compared to
the same table in the source, when we are doing Updates in the source
and an index is present in the table, the Updates fail on the replica
as the row to be changed is not found. Same applies to Deletes.

Analysis / Root-cause analysis
------------------------------

When the replication applier uses indexes to locate rows for Updates
and Deletes, it requires that all index parts are included in the
event. When an InnoDB table has a PK, then the PK is automatically a
(hidden) part of all non-unique indexes. Therefore, if the replica
table has a PK, the event must include all columns for the PK in order
for the applier to use indexes.

If for a given table, the PK column only exists in the replica version
of the table, then replication wont have all the values to search the
index. In order for the applier to search an index, the event needs to
have values for all key parts, including hidden ones. Replication only
checked if the user defined key parts were present till now which is
insufficient.

Implemented solution
---------------------

Replication now checks if all key parts are present before trying to
use an index to search data.

Change-Id: I0fa933aceba3c0d3f9cd66373df39cd788dbc379