Bug#33758255 - An edge case for base BUG 32038406 FILESORT USED FOR ORDER BY ...DESC EVEN WHEN DESCENDING INDEX AVAILABLE AND USED
Issue:

A filesort operation is introduced for a ORDER BY clause even if the
value is constant. This happens when the WHERE clause has a 'BETWEEN
low and high' and 'low=high', which is equivalent to having
'column=low' and no sort required.

Fix:
If the low and high arguments to a 'BETWEEN' function are same,
convert it to a 'EQUAL' function.

Change-Id: I7997ca47041db92e74611fea54ae20b0bc4bbb4c