Bug#34162207 - UNION support is partially broken after merging WL#14344 to mysql-trunk
All queries where there are UNION DISTINCT followed by UNION ALL were
not offloaded on mysql-trunk. Queries are of the form
1. (SELECT ...) UNION DISTINCT (SELECT ...) UNION ALL (SELECT ...)
2. CREATE VIEW V1 AS (SELECT ...) UNION
         DISTINCT (SELECT ...) UNION ALL (SELECT ...)

Problem: In the queries with UNION ALL, STREAM Access Path did not find
the correct JOIN Pointer/ Query Block, and hence did not find the correct
compilation context, which lead to failing assertions.

Solution:
Since Union(all) is a SET operation and query_terms are used in
MYSQL to represent SET Operations, and it is available on mysql-trunk,
we use query_terms to resolve query_blocks for union children.
This solution introduces query_term in the TranslateState to facilitate
this solution and the solution to BUG#34162247. Traversing query_term
tree additionally helps improves the following:
1. Readibility: Less corner cases to check for
     when accessing query blocks in presence of Set Operations.
2. Maintanability: Fewer single use code paths necessary to obtain
     correct query blocks. This will be fully accomplished after
     closing BUG#34323639.
3. Speed: Traversing query term tree is faster than traversing
     AP tree in some cases, since AP trees can be nested very
     deep, whereas, query term trees nest only as much as
     maximum depth of set operations.
Conceptually, this solution also helps to resolve the following
challenges which arise in translating the access path with
query terms:
1. Find appropriate query term,
   when multiple query terms are nested in 1 Query Expression and
2. Find appropriate Union Child's Join pointer
   when we have nested query terms.

Change-Id: I69f1028a71eff8fd45ff55f73a0ee51273476e75