BUG#34250614 MYSQLTEST FAILS TO SHOW WARNINGS
When mysqltest fails to run a query it's supposed to display the error
and also any warnings returned by the MySQL Server. However, when
mysqltest fails to read the result of the query it has sent, the display
of warnings are always turned off. It seems like the original assumption
was that failure to read query result is always a fatal error and thus
the "SHOW WARNINGS" query can't be sent to the server because the state
of protocol is unknown. But in most cases it's just a normal error
returned by the server and in that case it's well known that no more
results have been sent from the server, this will cause the client
library to turn off the SERVER_MORE_RESULTS_EXISTS flag in
cli_safe_read_with_ok_complete().

The missing display of SHOW WARNINGS makes it harder to diagnose
problems in the test environment since only the final error is available.
Having also the warnings available will make it easier to see the
chain of events that led up to the error and thus the root cause can
hopefully be found.

Fix by turning on display of warnings after error when the client
library indicates that there are no more results pending. In case of
failure where communication to server has been lost (or otherwise
failed) it will continue to not show any warnings.

Testing of this behavior can be performed by running a query that fails
with an error and produces warnings. For example with the ndbcluster
plugin, it's not supported to create tables with forward slashes in the
table name and creating such a table will now yield below output when
mysqltest fail to run the query:
  mysqltest: At line 89: Query 'CREATE TABLE `d/1`.`t1` (a int)
    ENGINE=NDB' failed.
  ERROR 1005 (HY000): Can't create table 't1' (use SHOW WARNINGS
    for more info).

  Warnings from just before the error:
  Warning 1296 Got error 4118 'Parameter error in API call' from NDB
  Warning 1296 Failed to discover table 't1' from NDB
  Warning 1296 Got error 4118 'Parameter error in API call' from NDB

Change-Id: I969f6692e7031429865ee0f89b3827b87cf97da4