Bug#34059253 Issues with NDB_STD_OPTS
  * Move --core-file option handling into ndbd / ndbmtd
  * Remove NDB_STD_OPTS; as a replacement, define each
    standard option individually in namespace NdbStdOpt
  * Remove '-c' as a short form for --connect-string
  * Remove description string from synonym options
    --ndb-mgmd-host and --connect-string
  * Use C++ linkage in ndb_opts
  * Fix Ndb_opts::set_usage_funcs() not to overwrite a pointer
    to the default function with a supplied null pointer
  * Fix some copy-and-paste errors in option definitions

This patch results in the options --charsets-dir and
--optimized-node-selection being removed from many utilities
that accepted but did not implement them.

While working on this I noted that mysql provides some more
modern facilities for option handling in client/base/, e.g.
abstract_program.h. In the future we could consider making
use of this in NDB.

This patch also in some cases results in a change in ordering
in the output of --help. I considered sorting options alpha-
betically in ndb_usage(); this might be done in a subsequent
patch.

Change-Id: Ib890f3dbe8412eb4ff95eb02e85a26b21d48a764