Bug#34258782 Bootstrapping with --name adds [DEFAULT].name to the config file.
When the user bootstraps the Router the configuration file gets created.
It contains options that depend on the command line parameters provided
by bootstrap command. Some of those options are not known to the config
file checker that verifies if unknown options are not present in the
config file.

This patch adds all supported options to the checker verification list.
It also adds a static assert mechanism to ensure each default option
added by the config generator is known by the checker.

Change-Id: I8e7f30373f3448e6e6aabfdf9dba401571ef675c