BUG#33176377: Contribution by Facebook: Fix race between binlog sender heartbeat timeout
Problem
=======
When the binlog sender thread is waiting for binlog
updates with heartbeat, it sporadically misses
update signals. In consequence, changes are not replicated
until the next signal is issued and noticed by the
binlog sender thread. The problem is waiting on conditional
variable with a timeout. When timeout in the binlog sender
has been reached, the thread needs to acquire ownership of
the mutex. At the same time, update signal may be issued
and mutex unlocked. Binlog sender takes ownership of the
mutex and checks if a timeout has been reached
- it is, so it assumes that signal has not been issued
and continues the loop, sends a heartbeat message and
loops again to wait on the conditional variable.

Solution
========
Check whether binlog has been updated after exitting with
a timeout from the mysql_cond_timedwait in the
wait_for_update function.
That kind of check is required after each timed wait on
a conditional variable. With implemented solution
wait_new_events is protected against spurious wake-ups.

Signed-off-by: Karolina Szczepankiewicz <karolina.szczepankiewicz@oracle.com>
Change-Id: I5c8bf874067ec301c6c1cf49cf7f2434f22d7e06