Bug#33801312: order matters for comparison between utf8mb4 and utf8mb3
Coercion of collations for two string values that are compared is
dependent upon order of operands in the comparison clause.
Given one string with utf8mb4 character set and one string with utf8mb3
character set, the comparison is carried out using the utf8mb4 collation
when the left operand is utf8mb4 and utf8mb3 when the left operand is
utf8mb3.

The problem is with special coercion of ASCII strings. Since both
strings are literals, they are scanned for ASCII and non-ASCII
characters, and the repertoire property is set according to this.
Notice then that the string 'ss' is ASCII, while 'ß' is non-ASCII,
thus the utf8mb4 string is considered to be ASCII and the utf8mb3
string is not. The code is then making a bad decision that the
utf8mb4 string is ASCII and thus compatible with the utf8mb3 string
and is converted before the comparison.

Fix: Use the MY_CS_PUREASCII property rather than MY_REPERTOIRE_ASCII.
MY_CS_PUREASCII is derived strictly from the character set, and not
the actual string, thus decision of collation becomes deterministic.

Because none of the strings are now identified as ASCII, the code that
identifies the utf8mb3 string as a true subset of the utf8mb4 is run,
and the utf8mb3 string is converted to utf8mb4 before the comparison.

No test case needed, for verification see the one for bug#34274525.

Change-Id: Iff61c3f5515ff130fadd691f3647fad54f58891a