Bug#34190101: Hypergraph Assert in FindBestQueryPlan on derived_condition_pushdown
The main.derived_condition_pushdown gets an assertion failure in a
query against information_schema tables when it's run with the
hypergraph optimizer.

The assertion failure is raised because two subplans for the same set
of tables are found to have different sets of active functional
dependencies. The functional dependencies are different because one of
the subplans has lost track of one of the predicates, and it has
therefore also lost track of the functional dependency associated with
that predicate.

The subplan that loses track of a predicate, is a sort-ahead on top of
a nested loop join. Because one of the inner tables of the nested loop
join is a lateral/correlated derived table with a BLOB column, a
materialization step has to be added between the sort-ahead path and
the nested loop join path, so that Filesort is given stable row IDs.

The problem is that the nested loop join path has a "delayed
predicate", which is a predicate that refers to one table in the
subplan, but also to one table that is not part of the subplan, and
the predicate therefore has to be applied higher up in the join tree
when the missing table is joined with this subplan. However,
CreateMaterializationPath() does not copy the delayed_predicates
member into the materialization path, and this predicate is therefore
not applied higher up in the join tree when this subplan is
considered.

The fix is to make CreateMaterializationPath() copy delayed_predicates
from the input path to the materialization path, so that the predicate
and the corresponding functional dependency are not lost.

Change-Id: Ia1a082d4bf4b7ca73a9229e7e8a2060bede22fb6