wl#14293 Add libmysql like Query Service in Server
At present we use the command service (WL#7947) to query locally within the
server. This service is tightly coupled with Protocol_classic. That means in order
to use the command service you have to provide the Protocol_callbacks. It
inconvenient to use besides that it exposes the protocol internals to user. It
creates more dependencies with the user code  hence make it harder to improve the
internals.

The users are already familiar with the libmysql C APIs which are well documented
and used by the clients (https://dev.mysql.com/doc/c-api/8.0/en/c-api-function-
overview.html).

"
Aim of this worklog is to either enable those client APIs to use within the server
or provide equivalent to them.
"

We evaluated both of these possibilities. We found that second option to be
simpler and easy to extend i.e. a component service.
We introduce a local query service. It will provide methods that could be used to
query within the server.

Advantages:

1. Easy to use, already familiar and extendable interface.
2. User need not to know or implement the internal details like protocols.
3. Components and Plugins can also use them.

This worklog is identified as one of the subtask of another larger goal that is to
have single port for different types of clients to communicate with server.
(WL#13894)

Approved by: Georgi Kodinov <georgi.kodinov@oracle.com>
Approved by: Bharathy Satish <bharathy.x.satish@oracle.com>