Bug#34298238: Assertion `cond->is_bool_func()' failed in MySQL 8.0.29
After a condition is pushed down to a derived table having a set
operation, while folding an always true boolean condition, the
re-write is not correct resulting in an assert. The wrong re-write is a
result of not telling the cloned condition to treat UNKNOWN as FALSE
when a copy is made during condition pushdown to derived tables with set
operations.
Solution is to set "abort_on_null" to true for the copied condition
which triggers the correct re-write during constant folding.

Change-Id: Id3c13adc43af46f867b2a0667e5aa7190477279a