Bug#25118903: MAX_JOIN_SIZE DOES NOT MATCH ROW ESTIMATES [2/2, max_join_size]
The max_join_size system variable is documented as a maximum number of
rows or disk seeks. The implementation, however, does not check number
of rows or disk seeks directly, but treats max_join_size as the
maximum estimated cost to permit. While cost and row count are
correlated, they are not the same.

This patch changes the implementation so that max_join_size is used to
limit the maximum number of row accesses in base tables. If the
estimates indicate that a higher number of rows will need to be read
from the base tables, an error is raised. This makes the actual
behaviour more in line with how it is documented.

The old check for max_join_size was performed only by the old
optimizer. The new implementation is moved into code that is common to
the old optimizer and the hypergraph optimizer, so that it takes
effect regardless of which optimizer is used.

Change-Id: Iabf1cca72f5d21f7622f237922776ba315798594