Bug#34306361: Assertion `in_item' failed in MySQL 8.0.29
When pushing a condition down to derived table with
set operations, a copy of the original condition is
made. This copy is used when replacing columns
in the condition with derived table expressions.
After the replacement, we update the used table
information and use this condition when processing
the query block. However, when multiple query blocks
are present, if the data types of the expressions
that are used to replace, differ in those query blocks,
we also need to set the correct types in the condition.
This is currently not being done. This results in an
assert during execution. For the query in bugpage,
the IN operator retains the result type of the left
operand as STRING_RESULT which is derived from the
first query block for the UNION operator. However
for the second query block, it needs to be INT_RESULT.
This mismatch leads to the assert.

Solution is to not push a condition down to a derived
table having set operations, if it has expressions
that have different result types across query blocks.

Change-Id: I3eb416ba0737dc7fb9c24c2be4e954de9cf6f701